+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/processor/ProcessorSupplierWithService.java
@@ -55,7 +55,13 @@ public final class ProcessorSupplierWithService<C, S> implements ProcessorSuppli
         HazelcastInstanceImpl hazelcastInstance = (HazelcastInstanceImpl) context.jetInstance().getHazelcastInstance();
         ManagedContext managedContext = hazelcastInstance.getSerializationService().getManagedContext();
         serviceContext = serviceFactory.createContextFn().apply(context);
-        serviceContext = (C) managedContext.initialize(serviceContext);
+        Object initializedObject = managedContext.initialize(serviceContext);
+        Class<?> serviceContextClass = serviceContext.getClass();
+        if (!serviceContextClass.isInstance(initializedObject)) {
+            throw new IllegalArgumentException(String.format("The initialized service context object should " +
+                    "be an instance of %s", serviceContextClass));
+        }
+        serviceContext = (C) initializedObject;
     }
 
     @Nonnull @Override
