+++ hazelcast-jet-spring/src/test/resources/com/hazelcast/jet/spring/application-context-processor.xml
@@ -44,5 +44,5 @@
 
     <jet:map instance-ref="jet-instance" name="my-map" id="my-map-bean"/>
 
-    <bean name="risk-calculator" class="com.hazelcast.jet.spring.SillyRiskCalculator" />
+    <bean name="risk-calculator" class="com.hazelcast.jet.spring.SillyCalculator" />
 </beans>
